refactor: move serialization code
