+++ tests/workflow/power_simulation.py
@@ -10,7 +10,9 @@ from dessia_common.models.workflows import simulation_workflow
 from dessia_common.models.power_test import components1, component_connections1, usage1
 
 
-workflow_run = simulation_workflow.run({0: components1, 1:component_connections1, 3:usage1})
+workflow_run = simulation_workflow.run({0: components1,
+                                        1:component_connections1,
+                                        3:usage1})
 print(workflow_run.log)
 
 workflow_run2 = DessiaObject.dict_to_object(workflow_run.to_dict())
@@ -28,7 +30,7 @@ evaluated = manual_run.block_evaluation(1)
 print(evaluated)
 assert not evaluated
 
-manual_run.add_input_value(3, usage)
+manual_run.add_input_value(3, usage1)
 
 evaluated_blocks = manual_run.continue_run()
 assert(manual_run.progress == 1)
